Update SpongeGradle to 0.8.2
SpongeGradle 0.8.1 contains a bug that will sometimes not write the metadata
to the `mcmod.info` if the class files are up to date. This causes the Sponge
API version not to be visible on Ore.

I'd also recommended explicitly putting a dependency on SpongeForge in your
@Plugin, as the project will be given a SpongeForge tag. This will better
communicate to users what platform your plugin can be used on.